refactor: move endpoints to home controller
